FIX: review comments
