+++ ton_client/client/src/json_interface/modules.rs
@@ -25,8 +25,8 @@ fn register_client(handlers: &mut RuntimeHandlers) {
     module.register_type::<crate::error::ClientError>();
     module.register_type::<crate::client::ClientConfig>();
     module.register_type::<crate::net::NetworkConfig>();
-    module.register_type::<crate::client::CryptoConfig>();
-    module.register_type::<ton_sdk::AbiConfig>();
+    module.register_type::<crate::crypto::CryptoConfig>();
+    module.register_type::<crate::abi::AbiConfig>();
 
     module.register_sync_fn_without_args(
         crate::client::get_api_reference,
