+++ ton_sdk/src/contract.rs
@@ -11,10 +11,10 @@
 * limitations under the License.
 */
 
-use crate::json_helper;
+use crate::{json_helper, SdkAbiConfig};
 use crate::local_tvm;
 use crate::error::SdkError;
-use crate::{AbiContract, BlockId, Message, MessageId, AbiConfig, Transaction};
+use crate::{AbiContract, BlockId, Message, MessageId, Transaction};
 
 use ed25519_dalek::{Keypair, PublicKey};
 use chrono::prelude::Utc;
@@ -915,7 +915,7 @@ impl Contract {
 
     // Add `expire` parameter to contract functions header
     fn make_expire_header(
-        abi_config: &AbiConfig,
+        abi_config: &SdkAbiConfig,
         abi: String,
         header: Option<String>,
         try_index: Option<u8>,
@@ -960,7 +960,7 @@ impl Contract {
         params: FunctionCallSet,
         internal: bool,
         key_pair: Option<&Keypair>,
-        timeouts: &AbiConfig,
+        timeouts: &SdkAbiConfig,
         try_index: Option<u8>,
     ) -> Result<SdkMessage> {
         let (header, expire) = Self::make_expire_header(timeouts, params.abi.clone(), params.header, try_index)?;
@@ -987,7 +987,7 @@ impl Contract {
     pub fn get_call_message_bytes_for_signing(
         address: MsgAddressInt,
         params: FunctionCallSet,
-        timeouts: &AbiConfig,
+        timeouts: &SdkAbiConfig,
         try_index: Option<u8>,
     ) -> Result<MessageToSign> {
         let (header, expire) = Self::make_expire_header(timeouts, params.abi.clone(), params.header, try_index)?;
@@ -1014,7 +1014,7 @@ impl Contract {
         image: ContractImage,
         key_pair: Option<&Keypair>,
         workchain_id: i32,
-        timeouts: &AbiConfig,
+        timeouts: &SdkAbiConfig,
         try_index: Option<u8>,
     ) -> Result<SdkMessage> {
         let (header, expire) = Self::make_expire_header(timeouts, params.abi.clone(), params.header, try_index)?;
@@ -1065,7 +1065,7 @@ impl Contract {
         params: FunctionCallSet,
         image: ContractImage,
         workchain_id: i32,
-        timeouts: &AbiConfig,
+        timeouts: &SdkAbiConfig,
         try_index: Option<u8>,
     ) -> Result<MessageToSign> {
         let (header, expire) = Self::make_expire_header(timeouts, params.abi.clone(), params.header, try_index)?;
